move the import of the updates in the v2 and ignore the v1 for now
